% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{ffespn_api}
\alias{ffespn_api}
\title{ESPN Fantasy Football API}
\usage{
ffespn_api(path, query = NULL, headers = NULL)
}
\arguments{
\item{path}{path to pull}

\item{query}{query parameters}

\item{headers}{headers}
}
\description{
ESPN Fantasy Football API
}
\keyword{internal}
