% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.r
\name{ffespn_projections}
\alias{ffespn_projections}
\title{Projections}
\usage{
ffespn_projections(
  season,
  week,
  pos = slot_names,
  scoring = c("ppr", "non_ppr"),
  league_id = getOption("ffespn.league_id", "942845238"),
  max_scoring_period = 2
)
}
\arguments{
\item{season}{integer year}

\item{week}{integer week (0 - 18)}

\item{pos}{character position. Ex. "QB", "RB", "RB/WR", "DST", "FLEX", "DT", ...}

\item{scoring}{scoring type}

\item{league_id}{custom league ID; defaults to option \code{ffespn.league_id}, then 942845238}

\item{max_scoring_period}{custom max_scoring_period (optional)}
}
\description{
Projections
}
