% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{slot_id_to_name}
\alias{slot_id_to_name}
\title{Slot Ids to Name}
\usage{
slot_id_to_name(x)
}
\arguments{
\item{x}{slot id}
}
\description{
Slot Ids to Name
}
\references{
{
\url{https://support.espn.com/hc/en-us/articles/115003939432-Roster-Slots-Offense-}
}
}
