% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.r
\name{ffespn_projections}
\alias{ffespn_projections}
\title{Projections}
\usage{
ffespn_projections(season, week, team)
}
\arguments{
\item{season}{integer year}

\item{week}{integer week (0 - 17)}

\item{team}{character team abbreviation}
}
\description{
Projections
}
