% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.r
\name{ffespn_projections}
\alias{ffespn_projections}
\title{Projections}
\usage{
ffespn_projections(season, week, pos = slot_names)
}
\arguments{
\item{season}{integer year}

\item{week}{integer week (0 - 17)}

\item{pos}{character position. Ex. "QB", "RB", "RB/WR", "DST", "FLEX", "DT", ...}
}
\description{
Projections
}
